+++ graphbrain/tests/test_corefs.py
@@ -2,11 +2,8 @@ import unittest
 
 from graphbrain import hedge
 from graphbrain import hgraph
-from graphbrain.corefs import are_corefs
-from graphbrain.corefs import coref_id
-from graphbrain.corefs import coref_set
-from graphbrain.corefs import main_coref
-from graphbrain.corefs import make_corefs
+from graphbrain.utils.corefs import (are_corefs, coref_id, coref_set,
+                                     main_coref, make_corefs)
 
 
 class TestCorefs(unittest.TestCase):